change the rsps average become more presision

+++ database/migrations/2019_01_30_094038_create_excels_table.php
@@ -26,7 +26,7 @@ class CreateExcelsTable extends Migration
             $table->decimal('prep_time',10,2)->nullable()->default(NULL);
             $table->decimal('travel_time',10,2)->nullable()->default(NULL);
             $table->decimal('work_time',10,2)->nullable()->default(NULL);
-            $table->decimal('rsps', 3, 2);
+            $table->integer('rsps');
             $table->timestamps();
         });
     }
